Refactor body creation out of simulation module
The simulation module has grown rather unwieldy.
