+++ .eslintrc.json
@@ -33,6 +33,7 @@
                 ".*\\.js",
                 ".*\\.json",
                 "@\\w*",
+                "#[0-9a-f]+",
                 "eslint-.*"
             ],
             "skipWords": [
